backend/ai-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
transformers==4.35.0
torch>=2.0.0
protobuf>=3.20.0
accelerate==0.24.1
sentencepiece==0.2.1
sacremoses==0.0.53

